requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.6
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
python-multipart==0.0.6
aiohttp==3.8.6
openai==1.3.0